pyproject.toml
[build-system]
requires = [
  "setuptools >= 64",
  "wheel",
  "setuptools_scm[toml]>=8",
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "celltransformer/_version.py"

[tool.setuptools]
packages = ["celltransformer"]

[tool.black]
target-version = ['py39', 'py310', 'py311']
skip-string-normalization = true
line-length = 79
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
  )
)
'''


[project]
name = "celltransformer"
authors = [{name = "Alex Lee", email = "alex.lee@ucsf.edu"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version"]
dependencies = [ 
  "pytest",
  "mkdocstrings[python] >= 0.18",
  "torch >= 2.1",
  "numpy >= 1.24.0",
  "scikit-learn >= 1.4.0",
	"omegaconf >= 2.0.0",
	"wandb >= 0.16.4",
	"hydra-core >=1.3.0",
	"pandas >= 1.5.0",
	"tqdm >= 4.0.0",
	"anndata >= 0.10.0",  
  "jaxtyping >= 0.2",
  "scvi-tools >= 1.0.0",
  "lightning >= 2.1.0",
  "colorcet",
]

[project.urls]
Home = "https//github.com/abbasilab/celltransformer"